+++ src/memoria.h
@@ -15,26 +15,26 @@
 using namespace std;
 
 //Implementacao da interface de memoria, para se conectar
-//com os modulos do RISC16
+//com os modulos do RISC-V
 struct mem : public sc_module, public mem_if {
 
-	int16_t read(const unsigned short adress);
+	int32_t read(const unsigned adress);
 
-	int16_t lw(const unsigned short address, int16_t constante);
+	int32_t lw(const unsigned address, int32_t constante);
 
-	void sw(const unsigned short address, int16_t constante, int16_t dado);
+	void sw(const unsigned address, int32_t constante, int32_t dado);
 
-	void write_mem(const unsigned short address, int16_t data);
+	void write_mem(const unsigned address, int32_t data);
 
 	void dump_mem(int inicio, int fim, char formato);
 
 
 	SC_CTOR(mem){
-		mem_ptr = new int16_t[MAX_MEM];
+		mem_ptr = new int32_t[MAX_MEM];
 	}
 
 private:
-		int16_t *mem_ptr;
+	int32_t *mem_ptr;
 };
